feat: FP-190 adds JIRA auto-issue GitHub template (#5)
* feat: FP-190 adds JIRA auto-issue GitHub template

Adds a GitHub issue template that automates the creation of
JIRA issues and corresponding Git branches.

This template provides a standardized form for users to
submit issues, which triggers a GitHub Action to create a
JIRA issue, update the GitHub issue title, and create a
Git branch based on the JIRA key.

* Removes auto issue template config

Removes the auto issue template configuration file.

This template is no longer required and is being removed to streamline the issue creation process.

Relates to FP-190